requirements.txt
blessed==1.14.2
certifi==2018.4.16
chardet==3.0.4
fuzzywuzzy==0.16.0
idna==2.6
python-Levenshtein==0.12.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
wcwidth==0.1.7
